requirements.txt
gym==0.21.0
numpy==1.19.2
ray[tune]
ray[rllib]
sortedcontainers==2.4.0
tensorboard==2.7.0
pandas
dm_tree
scipy
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
lz4
pyzmq
